Start of integration tests
This change adds a base integration class and a podman launcher that can
be used to test the api calls and their expected results. Currently the
podman calls that will work should be ones that can be performed as
rootless. Eventually calls that require privilege (e.g. network) can be
implemented but skipped if the tests are not invoked as root.  It should
be noted that all integration tests will be skipped if a podman
executable is not in the path.

Signed-off-by: Alex Schultz <aschultz@redhat.com>