api_connection: extract message from server response
Behavior prior to this commit:

In the api.request method, the error message returned by the server (for a request that generated an error 500, Internal Server Error) is not captured.
For example we might get:

    podman.errors.InternalServerError: Request POST:/v2.0.0/libpod/containers/create failed: Server got itself in trouble

Instead of:

    Decode(): json: cannot unmarshal string into Go struct field SpecGenerator.cni_networks of type []string

This makes troubleshooting issues in such requests difficult.

Behavior after this commit:

If the response body is in JSON and includes a "message" property, it is
extracted and included in the message of the exception raised.

closes issue #47

Signed-off-by: Nicolas Galler <ngl@odoo.com>