Remove env setup from Makefile
Besides being surprising to users/developers, doing this at runtime for
CI-testing is generally a bad idea.  Not only does it duplicate
operations often previously performed on the system. It also opens up
a big opportunity for automation-flakes to occur.

Fix this by removing the `env` Makefile target, and instead simply
documenting the necessary packages.

Signed-off-by: Chris Evich <cevich@redhat.com>