Fix image build method to handle containerfile location correctly
The docstring for BuildMixin.build() specifies that the dockerfile
argument is relative to the path/build context.

BuildMixin.build() calls api.tar_utils.prepare_containerfile() whose
docstring specifies that the return value is relative to the context
directory.

This commit fixes a bug where the absolute path to the containerfile
was being returned from prepare_containerfile() causing the subsequent
image build to fail as the containerfile was not found at the erroneous
path.

This commit also fixes an incorrect test for the prepare_containerfile()
function.