Correctly handle response from wait endpoint
The Podman REST wait endpoint returns a JSON encoded integer, whereas
the Container.wait() function was expecting a JSON
object with "Error" and "Status Code" keys (which actually corresponds
to the response from the "Podman compat" legacy API). So, we fix that
in the Container.wait() method, and in the consuming Container.run()
method.

See:
    https://docs.podman.io/en/latest/_static/api.html#tag/containers/operation/ContainerWaitLibpod
Signed-off-by: Jonathan G. Underwood <jonathan.underwood@gmail.com>