Includes the possibility to use the 'listTags' parameter in the images search operation.
From Podman 4.0, the endpoint for image search allows to include a
`listTags` (bool) parameter that allows the user get all the available
tags for a specific image in the repo.

The parameter is only included in the API request if the user includes
it in the kwargs parameters for the function.

Signed-off-by: apozsuse <andres.munoz@suse.com>
Signed-off-by: Andres Pozo Munoz <andres.munoz@suse.com>