Implementing the functionality of the 'named' argument
In the Image.save method, the named argument was ignored.
The self.id argument was always used, resulting in an empty repositories
file in the tarball.
As a result, if the tarball was loaded into Podman,
the image was not given a name (only None).
This limitation can cause problems in some cases,
since the name of the imported image is not known.

Signed-off-by: Milan Balazs <milanbalazs01@gmail.com>