Fix the locally non-existent image fails with AttributeError
The code for running a new container with an image that is not
present locally tries to access the ImagesManager
via self.client.images.
As self.client is an APIClient instance and not an instance of
PodmanClient, an AttributeError occurs:

AttributeError: 'APIClient' object has no attribute 'images'

With this fix the 'PodmanClient' object is used instead of
'APIClinet'.

Ticket:
https://github.com/containers/podman-py/issues/378

Signed-off-by: Milan Balazs <milanbalazs01@gmail.com>