Bump release to 5.4.0.1
The podman pyproject.toml was updated to fix a packaging issue that
prevented the `import podman.api` to work.

This caused an issue in the package versioning. Python-podman cannot be
bumped to version 5.4.1 since it will try to use the libpod/v5.4.1/
endpoint which is not released yet. Python-podman uses the same
versioning scheme and so the package need to add a fourth digi in the
version scheme to be released on PyPI. This can be avoided in
distribution packages with the option of a new release or a patch
downstream, but on GitHub and PyPI this is the workaround.

Source for packaging Python software
https://packaging.python.org/en/latest/specifications/version-specifiers/#version-specifiers

Signed-off-by: Nicola Sella <nsella@redhat.com>