Improve testing against distro and podman-next
1. Split upstream into distro/pnext test runs

A new CI run targets tests that are exclusive for podman-next.
/upstream plans are now split into /distro and /pnext to be explicit
against which podman version is used in the tests. Also, python version
used are now called base_python and all_python, to be more self
explainatory.

2. Add pytest marker `pnext` to filter tests

Tests that should run against podman-next are now skipped by default in
tox runs. They can be enabled with --pnext and filtered in using
`-m pnext`. These two options should be used together since `pnext` and
`non-pnext` tests are thought to be incompatible.

3. Split `test_container_mounts` to test a breaking change and update
   docs

The scenario `test_container_mounts` was split to demonstrates the
usage of the marker against an upstream breaking change.
CONTRIBUTING.md now reflects the new feature and explains how to use tox
locally to leverage on this feature.

4. Add manual trigger on packit

Since the tests cover a corner case and serve the purpose of testing
unreleased features of libpod (which is slightly out of the scope of
podman-py) the testing jobs are optional and should run for certain PRs
only. The command to run the pnext tests is the following.

/packit test --labels {pnext,podman_next}

Fixes: https://github.com/containers/podman-py/issues/547

Signed-off-by: Nicola Sella <nsella@redhat.com>