fix image build not using cache
This commit fixes issue #528 by adding a default value to parameters
layers and outputformat. This change aligns the behavior with
podman-remote. Function documentation is updated accordingly.
A new integration test (test_images.py:test_build_cache) checks that
caching is enabled and also that disabling it produces a different
output.

Signed-off-by: Federico Rizzo <fulminemizzega@yahoo.it>