Remove GH release from pypi workflow
The step to create GH release is clunky, confusing, and fails most of
the time. The code was supposed to be triggered on a new tag, but it's
simpler to create a release from the GH dashboard and let it create the
tag as well, which triggers pypi.

The workflow now mirrors the one from
packaging.python.org page

https://packaging.python.org/en/latest/guides/
  publishing-package-distribution-releases-using-
  github-actions-ci-cd-workflows/#the-whole-ci-cd-workflow

Signed-off-by: Nicola Sella <nsella@redhat.com>